Disable draw2d.enableAutoscale in Maven tests and adjust for scale in Tests
- Disabling it in the pom will take care of running tests from Maven
- For PDE Unit tests on Windows set -Ddraw2d.enableAutoscale=false
- ImageExportProviderTests needs to get display scale as float as it can be 1.25, 1.5 or 1.75

+++ tests/com.archimatetool.editor.tests/src/com/archimatetool/editor/diagram/ImageExportProviderTests.java
@@ -90,7 +90,7 @@ public class ImageExportProviderTests {
         assertTrue(tmp.exists());
 
         ImageData imageData = new ImageData(tmp.getPath());
-        int scale = ImageFactory.getImageDeviceZoom() / 100; // Image Exporter exports using this
+        float scale = ImageFactory.getImageDeviceZoom() / 100f; // Image Exporter exports using this
         assertEquals(148 * scale, imageData.width); // 128 width + 10 margins
         assertEquals(72 * scale, imageData.height); // 52 height + 10 margins
